Move query, getting templates to util.
