+++ detect_supernova/util.py
@@ -2,10 +2,10 @@ from dataclasses import dataclass
 import os
 from pathlib import Path
 import re
+import requests
 
 import pandas as pd
 
-from detect_supernova.make_openuniverse_subtraction_pairs import get_earliest_template_for_image
 from snappl.image import OpenUniverse2024FITSImage
 
 
@@ -36,6 +36,100 @@ class ImageInfo:
         self.skysub_path = self.temp_dir / f"skysub_{self.image_name}"
         self.detmask_path = self.temp_dir / f"detmask_{self.image_name}"
         self.psf_path = self.temp_dir / f"psf_{self.image_name}"
+
+
+def get_image_info_for_ra_dec(ra, dec, band=None):
+    server_url = "https://roman-desc-simdex.lbl.gov"
+    req = requests.Session()
+    json = {"containing": [ra, dec]}
+    if band is not None:
+        json["filter"] = band
+    result = req.post(f"{server_url}/findromanimages", json=json)
+    if result.status_code != 200:
+        raise RuntimeError(f"Got status code {result.status_code}\n{result.text}")
+    df = pd.DataFrame(result.json())
+    return df
+
+
+def get_templates_for_points(points, band, min_points=3):
+    """Returns all images in the same bandpass that overlap at least min_points
+    out of the list of points passed in.
+
+    Parameters
+    ----------
+    points : List of tuples of (ra, dec) points
+    band : str
+    min_points : int, number of required points images must cover
+
+    Returns
+    -------
+    images : List of DataFrames of image info from Roman-DESC-simdex
+
+    Notes
+    -----
+    This uses the roman-desc-simdex NERSC Spin server.
+    The DB server query for image info as a function of RA, Dec sometimes fails.
+    If it does then it will raise a RuntimeError, this function will let that pass through.
+    """
+    matches = []
+    for i, (ra, dec) in enumerate(points):
+        matching_images = get_image_info_for_ra_dec(ra, dec, band=band)
+        matches.append(matching_images)
+
+    matches = pd.concat(matches)
+    # From
+    #  https://stackoverflow.com/questions/35584085/how-to-count-duplicate-rows-in-pandas-dataframe
+    matches = matches.groupby(matches.columns.tolist()).size().reset_index().rename(columns={0: "counts"})
+    good_matches = matches.loc[matches.counts >= min_points]
+
+    return good_matches
+
+
+def get_templates_for_image(im, min_points=3):
+    """Return a list of matching images that could be used as templates.
+
+    Returns all images in the same bandpass that overlap at least min_points
+    out of the 5 points of the center + corners of the images
+
+    Parameters
+    ----------
+    images: Object with data attributes
+    ("ra", "dec", "ra_00", "dec_00", "ra_01", "dec_01", "ra_10", "dec_10", "ra_11", "dec_11")
+    and get method for "filter"
+    min_points: int
+
+    Returns
+    -------
+    images : list of (pointing, sca, band) tuples of overlapping images
+    """
+    corners = [
+        (im.ra_00, im.dec_00),
+        (im.ra_01, im.dec_01),
+        (im.ra_10, im.dec_10),
+        (im.ra_11, im.dec_11),
+    ]
+    center = [(im.ra, im.dec)]
+    points = center + corners
+
+    # band.filter would be a method so we can't use data attribute of same name
+    # and instead use `get` to access
+    band = im.get("filter")
+
+    return get_templates_for_points(points, band=band, min_points=min_points)
+
+
+def get_earliest_template_for_image(image, **kwargs):
+    """Get the earliest template that overlaps at least half the image.
+
+    Parameters
+    ----------
+    image : DataFrame of image info from Roman-DESC-simdex
+    """
+    templates = get_templates_for_image(image, **kwargs)
+    # Get earliest MJD
+    earliest_template = templates.iloc[templates.mjd.argsort()].iloc[0]
+
+    return earliest_template
 
 
 def get_center_and_corners(image_path):
